Thu, May 18, 2023, 4:49 PM +03:00

+++ finance/app.py
@@ -73,7 +73,7 @@ def buy():
 
         stocks = stock["name"]
         price = stock["price"]
-        total = price * shares
+        total = float(price * shares)
 
 
         current_user = session["user_id"]
